pyproject.toml
[project]
name = "anishot"
version = "0.1.0"
description = ""
authors = [
    {name = "externref",email = "shiva02939@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.9.7"
isort = "^6.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
